requirements.txt
streamlit==1.28.1
pandas==2.1.1
numpy==1.25.2
plotly==5.17.0
seaborn==0.12.2
matplotlib==3.7.2
openpyxl==3.1.2
xlrd==2.0.1
scikit-learn==1.3.0
scipy==1.11.3
statsmodels==0.14.0
groq==0.9.0
python-dotenv==1.0.0
streamlit-aggrid==0.3.4
streamlit-plotly-events==0.0.6
